requirements.txt
cmake
scikit-build
numpy
pillow
opencv-python==4.2.0.32
numpy
requests
flask
